Move step to buy one train of type per OR to game core
